feat: show boilerplate landing page for windows
This commit adds boilerplate `Landing` screen. This screen is currently
only being shown to Windows users.
