+++ src/setup/DockerNotDetected.test.tsx
@@ -6,6 +6,7 @@ test("renders 'Waiting for XUD Docker' text when promise is pending", () => {
   window.electron = {
     send: () => {},
     receive: () => {},
+    platform: () => "linux",
   };
   const { getByText } = render(<App />);
   const headerTextElement = getByText(/Waiting for XUD Docker/i);
